Add files via upload
Indicator lights are traffic control devices that assign the right-of-way to vehicle drivers and pedestrians and control their movements and turns by means of light and color signals that change over time at intersections or other special locations. Three colors are used as the main signal, assisted by sound. Signal lights in the city is defined as strict order, along with the development of urban traffic constantly change and improve. In the future world, when the human world has a better way to no longer experience traffic accidents, the indicator light can no longer play a warning role as a purely colorful light, given a definition by music that can appear simultaneously. Indicator lights can have a new definition and appear in a new form.